test(input): split scoped routing matrix
